requirements.txt
aiohttp==1.0.5
astroid==1.6.5
async-timeout==3.0.0
chardet==3.0.4
discord.py==0.16.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.3.1
mysql-connector-python==8.0.11
pep8==1.7.1
pkg-resources==0.0.0
protobuf==3.6.0
pylint==1.9.2
six==1.11.0
websockets==3.4
wrapt==1.10.11
